Refactor logging code
