+++ packages/devtools_app/lib/src/screens/logging/shared/constants.dart
@@ -2,4 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:intl/intl.dart';
+
 const loggingMinVerboseWidth = 650.0;
+
+final loggingTableTimeFormat = DateFormat('HH:mm:ss.SSS');
